+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 [patch]
-
+- Clean up unit tests 
 
 ## [1.3.8] - 2024-03-02
 - Add build pipeline